+++ src/app/pages/funcionarios/funcionario-grid/funcionario-grid.component.ts
@@ -1,7 +1,8 @@
-import { Component } from '@angular/core';
+import { Component, Inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToolboxService } from '../../../components/toolbox/toolbox.service';
 import { FuncionariosService } from '../../../services/funcionarios.service';
+import { MAT_DIALOG_DATA, MatDialog, MatDialogRef } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-funcionario-grid',
@@ -13,7 +14,9 @@ export class FuncionarioGridComponent {
   dataSource:any = [];
   dataSourceFilter:any = [];
   searchTerm: string = '';
-  constructor(private router: Router, private toolboxService: ToolboxService, private funcionariosService: FuncionariosService) {}
+  constructor(private router: Router, private toolboxService: ToolboxService, private funcionariosService: FuncionariosService,
+    public dialog: MatDialog
+  ) {}
 
   ngOnInit(): void {
     this.findAll();
@@ -48,7 +51,37 @@ export class FuncionarioGridComponent {
   }
 
   deleteItem(element: any){
-    this.funcionariosService.deleteItem(element.id);
-    this.findAll();
+    console.log(element)
+    const dialogRef = this.dialog.open(DialogDeleteFuncionario, {
+      width: '300px',
+      data: element
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        this.funcionariosService.deleteItem(element.id);
+        this.findAll();
+      }
+    });
   }
+}
+
+
+
+
+@Component({
+  selector: 'dialog-delete',
+  templateUrl: 'dialog-delete.html'
+})
+export class DialogDeleteFuncionario {
+  constructor(
+    public dialogRef: MatDialogRef<DialogDeleteFuncionario>,
+    @Inject(MAT_DIALOG_DATA) public data: any) {}
+    onYesClick(): void {
+      this.dialogRef.close(true);
+    }
+
+    onCancelClick(): void {
+      this.dialogRef.close(false);
+    }
 }
