+++ src/app/pages/imoveis/imovel-grid/imovel-grid.component.ts
@@ -1,7 +1,8 @@
-import { Component } from '@angular/core';
+import { Component, Inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToolboxService } from '../../../components/toolbox/toolbox.service';
 import { ImoveisService } from '../../../services/imoveis.service';
+import { MAT_DIALOG_DATA, MatDialog, MatDialogRef } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-imovel-grid',
@@ -13,7 +14,7 @@ export class ImovelGridComponent {
   dataSource:any = [];
   dataSourceFilter:any = [];
   searchTerm: string = '';
-  constructor(private router: Router, private imoveisService: ImoveisService) {}
+  constructor(private router: Router, private imoveisService: ImoveisService, public dialog: MatDialog) {}
   addNew() {
     this.router.navigate(["/imovel/novo"]);
   }
@@ -47,7 +48,36 @@ export class ImovelGridComponent {
   }
 
   deleteItem(element: any){
-    this.imoveisService.deleteItem(element.id);
-    this.findAll();
+    console.log(element)
+    const dialogRef = this.dialog.open(DialogDelete, {
+      width: '300px',
+      data: element
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        this.imoveisService.deleteItem(element.id);
+        this.findAll();
+      }
+    });
   }
+}
+
+
+
+@Component({
+  selector: 'dialog-delete',
+  templateUrl: 'dialog-delete.html'
+})
+export class DialogDelete {
+  constructor(
+    public dialogRef: MatDialogRef<DialogDelete>,
+    @Inject(MAT_DIALOG_DATA) public data: any) {}
+    onYesClick(): void {
+      this.dialogRef.close(true);
+    }
+
+    onCancelClick(): void {
+      this.dialogRef.close(false);
+    }
 }
