+++ src/app/pages/vendedores/vendedor-grid/vendedor-grid.component.ts
@@ -1,7 +1,8 @@
-import { Component } from '@angular/core';
+import { Component, Inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToolboxService } from '../../../components/toolbox/toolbox.service';
 import { VendedoresService } from '../../../services/vendedores.service';
+import { MAT_DIALOG_DATA, MatDialog, MatDialogRef } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-vendedor-grid',
@@ -13,7 +14,9 @@ export class VendedorGridComponent {
   dataSource:any = [];
   dataSourceFilter:any = [];
   searchTerm: string = '';
-  constructor(private router: Router, private toolboxService: ToolboxService, private vendedoresService: VendedoresService) {}
+  constructor(private router: Router, private toolboxService: ToolboxService, private vendedoresService: VendedoresService,
+    public dialog: MatDialog
+  ) {}
 
  
   ngOnInit(): void {
@@ -49,7 +52,36 @@ export class VendedorGridComponent {
   }
 
   deleteItem(element: any){
-    this.vendedoresService.deleteItem(element.id);
-    this.findAll();
+    console.log(element)
+    const dialogRef = this.dialog.open(DialogDelete, {
+      width: '300px',
+      data: element
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        this.vendedoresService.deleteItem(element.id);
+        this.findAll();
+      }
+    });
   }
+}
+
+
+
+@Component({
+  selector: 'dialog-delete',
+  templateUrl: 'dialog-delete.html'
+})
+export class DialogDelete {
+  constructor(
+    public dialogRef: MatDialogRef<DialogDelete>,
+    @Inject(MAT_DIALOG_DATA) public data: any) {}
+    onYesClick(): void {
+      this.dialogRef.close(true);
+    }
+
+    onCancelClick(): void {
+      this.dialogRef.close(false);
+    }
 }
